bugfix. don't move the file if the destination is the same as the source improvement. make the links frame not stretch the entry widgets

+++ python_app/outside_calls/fsApp_calls.py
@@ -17,8 +17,9 @@ class FinderCalls:
 
     @classmethod
     def moveFile(cls, sourceFilepath, destFilepath):
-        cls.copyFile(sourceFilepath, destFilepath)
-        cls.deleteFile(sourceFilepath)
+        if sourceFilepath != destFilepath:
+            cls.copyFile(sourceFilepath, destFilepath)
+            cls.deleteFile(sourceFilepath)
 
     @classmethod
     def copyFile(cls, sourceFilepath, destFilepath):
